functions/requirements.txt
azure-functions>=1.20.0,<2.0.0
azure-cosmos>=4.7.0,<5.0.0
azure-eventgrid>=4.20.0,<5.0.0
azure-identity>=1.17.1,<2.0.0
